Refactor classes - move classes for test to another package Add using DBUtils for other tests Remove unused part of method clearDatabase() in DBUtils Fixed bugs: getTypeOfPrimaryKeyFieldSQL - not correctly work with UUID - to early exception setValueOfFieldForObject - not correctly set value if value was get null
